requirements.txt
aiohttp
asyncio
nextcord
nextcord.py[voice]
pyyaml
pandas
mysql
mysql-connector
dateparser
pytz
requests
aiofiles
inspiro
trafilatura
sklearn
nltk
schedule
apscheduler
uwuify
autocorrect
tensorflow
cmudict
contractions
language-tool-python
colour
geopy
Nominatim
haversine